+++ tests/bootstrap.js
@@ -4,7 +4,7 @@
 const mock = require('mock-require');
 
 mock('@janiscommerce/mt188-card-reader', {
-	MT188CardReader: {
-		getPANWithRetries: () => Promise.resolve({})
+	MT188CardReader: class MT188CardReader {
+		getPANWithRetries() { return Promise.resolve({}); }
 	}
 });